Created a new Component Folder
